Updated run.py version and loading

+++ Regression/teeks99-06/run.py
@@ -1,4 +1,4 @@
-#!/usr/bin/python
+#!/usr/bin/env python
 
 # Copyright Rene Rivera 2007-2013
 #
@@ -24,16 +24,6 @@ use_local = use_local_argument in sys.argv
 if use_local:
     del sys.argv[sys.argv.index(use_local_argument)]
 
-# Assume we use git.
-use_git = True
-
-# But allow overriding with svn for legacy comparison testing.
-use_svn_argument = '--use-svn'
-use_svn = use_svn_argument in sys.argv
-
-if use_svn:
-    use_git = False
-
 #~ The directory this file is in.
 if use_local:
     root = os.path.abspath(os.path.realpath(os.path.curdir))
@@ -42,14 +32,12 @@ else:
 print '# Running regressions in %s...' % root
 
 script_sources = [ 'collect_and_upload_logs.py', 'process_jam_log.py', 'regression.py' ]
-script_local = os.path.join(root,'tools','regression','src')
+script_local = root
 if use_local:
     script_remote = 'file://'+os.path.abspath(os.path.dirname(os.path.realpath(__file__)))
-elif use_git:
-    script_remote = 'https://raw.github.com/boostorg/boost/develop/tools/regression/src'
 else:
-    script_remote = 'http://svn.boost.org/svn/boost/trunk/tools/regression/src'
-script_dir = os.path.join(root,'tools_regression_src')
+    script_remote = 'https://raw.githubusercontent.com/boostorg/regression/develop/testing/src'
+script_dir = os.path.join(root,'boost_regression_src')
 
 if not no_update:
     #~ Bootstrap.
@@ -58,25 +46,25 @@ if not no_update:
     if os.path.exists(script_dir):
         shutil.rmtree(script_dir)
     os.mkdir(script_dir)
-    #~ * Get new scripts, either from local working copy, or from svn
-    if os.path.exists(script_local):
+    #~ * Get new scripts, either from local working copy, or from remote
+    if use_local and os.path.exists(script_local):
         print '# Copying regression scripts from %s...' % script_local
         for src in script_sources:
             shutil.copyfile( os.path.join(script_local,src), os.path.join(script_dir,src) )
     else:
-        print '# Dowloading regression scripts from %s...' % script_remote
+        print '# Downloading regression scripts from %s...' % script_remote
         proxy = None
         for a in sys.argv[1:]:
             if a.startswith('--proxy='):
-                proxy = {'http' : a.split('=')[1] }
-                print '--- %s' %(proxy['http'])
+                proxy = {'https' : a.split('=')[1] }
+                print '--- %s' %(proxy['https'])
                 break
         for src in script_sources:
             urllib.FancyURLopener(proxy).retrieve(
                 '%s/%s' % (script_remote,src), os.path.join(script_dir,src) )
 
-#~ * Make the scripts are available to Python
-sys.path.insert(0,os.path.join(root,'tools_regression_src'))
+#~ * Make the scripts available to Python
+sys.path.insert(0,os.path.join(root,'boost_regression_src'))
 
 #~ Launch runner.
 from regression import runner
